Require deletion confirmation annotation for Project resources
